Add instance checking for older versions of js_of_ocaml
Currently in js_of_ocaml master (post commit
ocsigen/js_of_ocaml@887507db1eb8efd779070cbedab3774098a52939) the int64
implementation is backed by an MlInt64 class. Previous to this commit,
the representation was instead an Array of length 4, of the format
`[255, lo, mi, hi]`.

This code now implements both versions, falling back to the old
representation when an int64 is represented as an Array.

Note: This will break compilation on newer versions, as the 'requires'
fields contain some values that are not present after the commit. This
commit adds a TODO for addressing this in the future.